Move CLI test for `inventory lint` to `test_cli_functions.py`
