+++ tests/test_cli_functions.py
@@ -1,7 +1,11 @@
+from pathlib import Path
+from typing import Any, Dict, List
 from unittest import mock
 
-from click.testing import CliRunner
 import pytest
+import yaml
+
+from click.testing import CliRunner
 
 from commodore import cli
 
@@ -22,7 +26,9 @@ from commodore import cli
     ],
 )
 @mock.patch.object(cli, "fetch_token")
-def test_commodore_fetch_token(fetch_token, capsys, args, exitcode, output):
+def test_commodore_fetch_token(
+    fetch_token, capsys, args: List[str], exitcode: int, output: str
+):
     fetch_token.side_effect = lambda cfg: "id-1234"
     runner = CliRunner()
 
@@ -30,3 +36,81 @@ def test_commodore_fetch_token(fetch_token, capsys, args, exitcode, output):
 
     assert result.exit_code == exitcode
     assert result.stdout == output
+
+
+@pytest.mark.parametrize(
+    "files,exitcode,stdout",
+    [
+        ({}, 0, ["No errors"]),
+        (
+            {
+                "test.yaml": {
+                    "parameters": {
+                        "components": {
+                            "tc1": {
+                                "url": "https://example.com/tc1.git",
+                                "version": "v1.0.0",
+                            },
+                            "tc2": {
+                                "url": "https://example.com/tc2.git",
+                                "version": "feat/test",
+                            },
+                            "tc3": {
+                                "url": "https://example.com/tc3.git",
+                                "version": "master",
+                            },
+                        },
+                        "customer_name": "${cluster:tenant}",
+                    }
+                }
+            },
+            0,
+            ["No errors"],
+        ),
+        (
+            {
+                "test.yaml": {
+                    "parameters": {
+                        "components": {
+                            "tc1": {
+                                "url": "https://example.com/tc1.git",
+                            },
+                            "tc2": {
+                                "url": "https://example.com/tc2.git",
+                                "version": "feat/test",
+                            },
+                            "tc3": {
+                                "url": "https://example.com/tc3.git",
+                                "version": "master",
+                            },
+                        },
+                        "customer_name": "${customer:name}",
+                    }
+                }
+            },
+            1,
+            [
+                "> Component specification for tc1 is missing explict version in {0}/test.yaml",
+                "> Field 'parameters.customer_name' in file '{0}/test.yaml' "
+                + "contains deprecated parameter '${{customer:name}}'",
+                "Found 2 errors",
+            ],
+        ),
+    ],
+)
+def test_inventory_lint_cli(
+    tmp_path: Path,
+    files: Dict[str, Dict[str, Any]],
+    exitcode: int,
+    stdout: List[str],
+):
+    runner = CliRunner()
+
+    for f, data in files.items():
+        with open(tmp_path / f, "w") as fh:
+            yaml.safe_dump(data, fh)
+
+    result = runner.invoke(cli.commodore, ["inventory", "lint", str(tmp_path)])
+
+    assert result.exit_code == exitcode
+    assert all(line.format(tmp_path) in result.stdout for line in stdout)
