--- tests/test_inventory_lint.py
@@ -1,88 +1,0 @@
-from pathlib import Path
-from typing import Any, Dict, Iterable
-from subprocess import run
-
-import pytest
-import yaml
-
-
-@pytest.mark.parametrize(
-    "files,exitcode,stdout",
-    [
-        ({}, 0, ["No errors"]),
-        (
-            {
-                "test.yaml": {
-                    "parameters": {
-                        "components": {
-                            "tc1": {
-                                "url": "https://example.com/tc1.git",
-                                "version": "v1.0.0",
-                            },
-                            "tc2": {
-                                "url": "https://example.com/tc2.git",
-                                "version": "feat/test",
-                            },
-                            "tc3": {
-                                "url": "https://example.com/tc3.git",
-                                "version": "master",
-                            },
-                        },
-                        "customer_name": "${cluster:tenant}",
-                    }
-                }
-            },
-            0,
-            ["No errors"],
-        ),
-        (
-            {
-                "test.yaml": {
-                    "parameters": {
-                        "components": {
-                            "tc1": {
-                                "url": "https://example.com/tc1.git",
-                            },
-                            "tc2": {
-                                "url": "https://example.com/tc2.git",
-                                "version": "feat/test",
-                            },
-                            "tc3": {
-                                "url": "https://example.com/tc3.git",
-                                "version": "master",
-                            },
-                        },
-                        "customer_name": "${customer:name}",
-                    }
-                }
-            },
-            1,
-            [
-                "> Component specification for tc1 is missing explict version in {0}/test.yaml",
-                "> Field 'parameters.customer_name' in file '{0}/test.yaml' "
-                + "contains deprecated parameter '${{customer:name}}'",
-                "Found 2 errors",
-            ],
-        ),
-    ],
-)
-def test_inventory_lint_cli(
-    tmp_path: Path,
-    files: Dict[str, Dict[str, Any]],
-    exitcode: int,
-    stdout: Iterable[str],
-):
-
-    for f, data in files.items():
-        with open(tmp_path / f, "w") as fh:
-            yaml.safe_dump(data, fh)
-
-    result = run(
-        f"commodore inventory lint {tmp_path}",
-        shell=True,
-        capture_output=True,
-    )
-
-    assert result.returncode == exitcode
-    out = result.stdout.decode("utf-8")
-    assert all(line.format(tmp_path) in out for line in stdout)
